Add `gtk/gtk.h` include
I add the includes to the `header.h` file rather than the
`module.modulemap` file because the modulemap file doesn’t search the
include paths for listed files, so I put `header.h` and that will
always be found. Inside the `header.h` file, the include directories
*are* searched when `#include` is used, so I can put normal includes
here and not worry about the exact path.